requirements.txt
torch
hydra-core
wandb
tqdm
scipy
numpy
matplotlib
pymunk
opencv-python
torch-geometric
torch-cluster
networkx